[FIX] Modify onboarding and estimated time return feature

+++ src/main/java/com/dh/ondot/schedule/api/response/AlarmDto.java
@@ -14,7 +14,7 @@ public record AlarmDto(
         Integer snoozeCount,
         String soundCategory,
         String ringTone,
-        Integer volume
+        Double volume
 ) {
     public static AlarmDto of(Alarm alarm) {
         return new AlarmDto(
